api call properly added
